Adjust qeueue.broken.c example
The problem with queue.broken.c is that IntQueueAux can't pack.
A solution to this is to `/*@ split_case h == c; @*/` but since
that creates another problem where CN tries to merge padding bytes
in the case h == c.

Therefore, this version ajdusts the mallocIntQueueCell() function
to stop the return value aliasing by construction.

The main edits were:
* Adjusting mallocIntQueueCell() function
* Tidying up the predicates by removing the dummy argument.
* Moving the assert about head & tail earlier.

It also adjusts IntQueue1 and IntQueueAux to use snoc. This is
experimental and may not work.