Record all output for tests
Cerberus repo has a test program, `diff-prog` which supports
parallelised testing with rich output capture, and an easy way to
update tests by emitting diffs which can be fed to patch.

diff-prog requires a JSON file for each configuration (one for Z3,
another for CVC5) and a bank of expected outputs, so this commit
adds both of those.