mkdocs.yml: Add GitHub widget to tutorial pages.
This makes sure there is a link to the GitHub-hosted tutorial repository on
every page (either visible next to the search on large screens, or in the
navigation drawer on small screens).

Someone has already added the setting necessary to display "Edit this page" and
"View this page" icons on each page -- this commit adds the requisite edit_uri
configuration, which we need as we don't use `master` as our trunk branch
name (note also that these links will be broken until the changes are merged
into main; there is no obvious way to have edit_uri pick up the current branch
for the benefit of local testing).