main.py: Use an environment variable to conditionally include TODOs.
main.py now checks whether or not an environment variable, TUTORIAL_TODOS, is
present in the environment (for now, the value is unimportant; this could be
changed to support e.g. different levels of commentary inclusion.)

README.md also includes instructions on using this variable to include developer
commentary when building/serving the HTML documentation.

This ought to give us the desired behavior in CI "for free", as the default
build now excludes the TODO commentary.