Add script (and docs) to assist in Python environment installation.
This is an attempt to make the development of CN tutorials more straightforward,
as far as setting up the Python build is concerned.

scripts/setup_env.sh handles creating a Python virtual environment suitable for
building the tutorial (that is, with the necessary dependencies).
This script tries to be relatively smart about not doing extra work.

The README has been updated to reflect these changes.